refactor: NaverBookSearchItem 모델을 별도 파일로 분리
