+++ app/src/main/java/com/boostcamp/and03/data/model/response/NaverBookSearchResponse.kt
@@ -15,11 +15,3 @@ data class NaverBookSearchResponse(
     val nextStart: Int
         get() = start + display
 }
-
-@Serializable
-data class NaverBookSearchItem(
-    val title: String,
-    val image: String,
-    val author: String,
-    val publisher: String,
-)
